feat: add useSecurityAutoSwitch
